\name{photonMonkey-package}
\alias{photonMonkey-package}
\alias{photonMonkey}
\docType{package}
\title{
Misc Tools for Photonics
}
\description{
Misc Tools for Photonics
}
\details{
\tabular{ll}{
Package: \tab photonMonkey\cr
Type: \tab Package\cr
Version: \tab 0.9\cr
Date: \tab 2012-10-17\cr
License: \tab GPL\cr
}

}
\author{
Tom Constant
}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}