\name{photonMonkey-package}
\alias{photonMonkey-package}
\alias{photonMonkey}
\docType{package}
\title{
Misc Tools for Photonics
}
\description{
Misc Tools for Photonics
}
\details{
\tabular{ll}{
Package: \tab photonMonkey\cr
Type: \tab Package\cr
Version: \tab 0.9\cr
Date: \tab 2012-10-17\cr
License: \tab GPL\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Tom Constant
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}